% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_15_min_data}
\alias{get_15_min_data}
\title{Get 15 minute interval data from fitbit using cookie returned from login function}
\usage{
get_15_min_data(cookie, what = "steps", date)
}
\arguments{
\item{cookie}{Cookie returned after login, specifically the "u" cookie}

\item{what}{What data you wish to be returned. Options include "steps", "distance". Others to come.}

\item{date}{Date in YYYY-MM-DD format}
}
\description{
Enter a string and have latitude and longitude returned using the HERE API
}
\examples{
\dontrun{
get_15_min_data(cookie, what="steps", "2015-01-20")
}
get_15_min_data
}
\keyword{data}

