% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_daily_data.R
\name{get_daily_data}
\alias{get_daily_data}
\title{Get daily data from fitbit.com}
\usage{
get_daily_data(cookie, what = "steps", start_date, end_date)
}
\arguments{
\item{cookie}{Cookie returned after login, specifically the "u" cookie}

\item{what}{What data you wish to be returned. Options include "steps", "distance", "floors", "minutesVery", "caloriesBurnedVsIntake", "getTimeInHeartRateZonesPerDay"}

\item{start_date}{Date in YYYY-MM-DD format}

\item{end_date}{Date in YYYY-MM-DD format}
}
\value{
A dataframe with two columns:
 \item{time}{A POSIXct time value}
 \item{data}{The data column corresponding to the choice of "what"}
}
\description{
Get daily data from fitbit using cookie returned from login function
}
\examples{
\dontrun{
get_daily_data(cookie, what="steps", start_date="2015-01-13", end_date="2015-01-20")
}
get_daily_data
}
\keyword{data}

