% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_intraday_data.R
\name{get_intraday_data}
\alias{get_intraday_data}
\title{Get intraday data from fitbit.com}
\usage{
get_intraday_data(cookie, what = "steps", date)
}
\arguments{
\item{cookie}{Cookie returned after login, specifically the "u" cookie}

\item{what}{What data you wish to be returned. Options include "steps", "distance", "floors", "active-minutes", "calories-burned", "heart-rate"}

\item{date}{Date in YYYY-MM-DD format}
}
\value{
A dataframe with two columns:
 \item{time}{A POSIXct time value}
 \item{data}{The data column corresponding to the choice of "what"}
}
\description{
Get intraday data from fitbit using cookie returned from login function
}
\examples{
\dontrun{
get_intraday_data(cookie, what="steps", date="2015-01-20")
}
get_intraday_data
}
\keyword{data}

