% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_sleep_data}
\alias{get_sleep_data}
\title{Get sleep data from fitbit.com}
\usage{
get_sleep_data(cookie, start_date = "2015-01-13", end_date = "2015-01-20")
}
\arguments{
\item{cookie}{Cookie returned after login, specifically the "u" cookie}

\item{start_date}{Date in YYYY-MM-DD format}

\item{end_date}{Date in YYYY-MM-DD format}
}
\description{
Get sleep data from fitbit using cookie returned from login function
}
\examples{
\dontrun{
get_sleep_data(cookie, start_date="2015-01-13", end_date="2015-01-20")
}
get_sleep_data
}
\keyword{data}

