% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_weight_data.R
\name{get_weight_data}
\alias{get_weight_data}
\title{Get weight data from fitbit.com}
\usage{
get_weight_data(cookie, start_date, end_date)
}
\arguments{
\item{cookie}{Cookie returned after login, specifically the "u" cookie}

\item{start_date}{Date in YYYY-MM-DD format}

\item{end_date}{Date in YYYY-MM-DD format}
}
\value{
A dataframe with two columns:
 \item{time}{A POSIXct time value}
 \item{weight}{The data column corresponding to weight}
}
\description{
Get weight data from fitbit using cookie returned from login function
}
\examples{
\dontrun{
get_weight_data(cookie, start_date="2015-01-13", end_date="2015-01-20")
}
get_weight_data
}
\keyword{data}

