% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_15_min_data.R
\name{get_15_min_data}
\alias{get_15_min_data}
\title{Get 15 minute interval data from fitbit.com}
\usage{
get_15_min_data(...)
}
\arguments{
\item{...}{Arguments from get_intraday_data()}
}
\value{
A dataframe with two columns:
 \item{time}{A POSIXct time value}
 \item{data}{The data column corresponding to the choice of "what"}
}
\description{
Get 15 minute interval data from fitbit using cookie returned from login function
}
\examples{
\dontrun{
get_15_min_data(cookie, what="steps", date="2015-01-20")
}
get_15_min_data
}
\keyword{data}

