% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_sleep_data.R
\name{get_sleep_data}
\alias{get_sleep_data}
\title{Get sleep data from fitbit.com}
\usage{
get_sleep_data(cookie, start_date = "2015-01-13", end_date = "2015-01-20")
}
\arguments{
\item{cookie}{Cookie returned after login, specifically the "u" cookie}

\item{start_date}{Date in YYYY-MM-DD format}

\item{end_date}{Date in YYYY-MM-DD format}
}
\value{
A list with two things
 \item{summary}{A list of sleep summary values}
 \item{df}{A data frame containing various sleep values over time}
}
\description{
Get sleep data from fitbit using cookie returned from login function
}
\examples{
\dontrun{
get_sleep_data(cookie, start_date="2015-01-13", end_date="2015-01-20")
}
get_sleep_data
}
\keyword{data}

