% Generated by roxygen2 (4.0.2): do not edit by hand
\name{login}
\alias{login}
\title{Login to fitbit.com and get cookie}
\usage{
login(email, password)
}
\arguments{
\item{email}{Email address used to login to fitbit.com}

\item{password}{Password used to login to fitbit.com}
}
\description{
Get the login cookie after login at www.fitbit.com
}
\examples{
\dontrun{
cookie <- login(email="corynissen<at>gmail.com", password="mypasswordhere")
}
login
}
\keyword{login}

